MLIR: fix MERGE reads and writes, split the merge executor
