+++ query/ir/dialect/nl/NLOps.td
@@ -511,7 +511,8 @@ def GetNodeProperties : NLOp<"get_node_properties", [Pure, RowAlignedThroughOper
   // buildable, so their types are omitted; only the nullable value chunk type is
   // spelled, after the colon. `pending` is the mask an nl.merge produced beside its
   // node chunk: such a row holds a write-buffer offset rather than an ID the graph
-  // knows, so its value reads null instead of the node the offset collides with.
+  // knows, so its value is read out of the write buffer rather than off the node the
+  // offset collides with - null there when the write set the property on none.
   let arguments = (ins NLNodeIDChunk:$input_nodes,
                        NLPropertyTypeHandle:$property_type,
                        Optional<NLBoolChunk>:$pending);
